+++ reana-ui/src/pages/signin/Signin.js
@@ -10,109 +10,78 @@
 
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { Button, Divider, Grid, Image, Segment } from "semantic-ui-react";
-import { Link, useHistory, useLocation } from "react-router-dom";
-import PropTypes from "prop-types";
+import { Button, Divider, Segment } from "semantic-ui-react";
+import { Link } from "react-router-dom";
 
 import { getConfig } from "~/selectors";
 import SignForm from "./components/SignForm";
+import SignContainer from "./components/SignContainer";
 import { api } from "~/config";
-import { userSignup, userSignin } from "~/actions";
-import LogoImg from "~/images/logo-reana.svg";
+import { triggerNotification, userSignin } from "~/actions";
+import { useSubmit } from "~/hooks";
 
-import styles from "./Signin.module.scss";
-
-export default function Signin({ signup }) {
+export default function Signin() {
+  const handleSubmit = useSubmit(userSignin);
   const config = useSelector(getConfig);
-  const [formData, setFormData] = useState({ email: "", password: "" });
   const dispatch = useDispatch();
-  const history = useHistory();
-  const location = useLocation();
+  const [formData, setFormData] = useState({ email: "", password: "" });
 
   const handleClick = () => {
     window.location.href = api + "/oauth/login/cern";
-  };
-
-  const handleSubmit = (event, action) => {
-    const { from } = location.state || { from: { pathname: "/" } };
-    dispatch(action(formData)).then((res) => {
-      if (res.isAxiosError ?? false) {
-        setFormData({ ...formData, password: "" });
-      } else {
-        history.replace(from);
-      }
-    });
-    event.preventDefault();
+    // FIXME: We assume that the sign-up went successfully but we actually don't know.
+    // We should upgrade Invenio-OAuthClient to latest version that supports REST apps
+    // and adapt the whole workflow.
+    if (config.userConfirmation) {
+      dispatch(
+        triggerNotification(
+          "Success!",
+          "User registered. Please confirm your email by clicking on the link we sent you."
+        )
+      );
+    }
   };
 
   const handleInputChange = (event) => {
-    const target = event.target;
+    const { target } = event;
     setFormData({ ...formData, [target.name]: target.value });
   };
 
   return (
-    <div className={styles["signin-form"]}>
-      <Grid
-        textAlign="center"
-        verticalAlign="middle"
-        className={styles["signin-grid"]}
-      >
-        <Grid.Column className={styles["signin-column"]}>
-          <Image
-            centered
-            spaced
-            src={LogoImg}
-            size="medium"
-            className={styles["reana-logo"]}
-          />
-          <Segment>
-            {config.cernSSO && !signup && (
-              <>
-                <Button basic fluid size="large" onClick={handleClick}>
-                  Sign in with SSO
-                </Button>
-                {config.localUsers && (
-                  <Divider section horizontal>
-                    or
-                  </Divider>
-                )}
-              </>
-            )}
+    <SignContainer>
+      <Segment>
+        {config.cernSSO && (
+          <>
+            <Button basic fluid size="large" onClick={handleClick}>
+              Sign in with SSO
+            </Button>
             {config.localUsers && (
-              <SignForm
-                submitText={signup ? "Sign up" : "Sign in"}
-                handleSubmit={(e) =>
-                  handleSubmit(e, signup ? userSignup : userSignin)
-                }
-                formData={formData}
-                handleInputChange={handleInputChange}
-              />
+              <Divider section horizontal>
+                or
+              </Divider>
             )}
-          </Segment>
-          {config.localUsers && (
-            <>
-              {!signup ? (
-                <p>
-                  Don't have an account? <Link to="/signup">Sign up</Link>
-                </p>
-              ) : (
-                <p>
-                  Already signed up? Go to <Link to="/signin">Sign In</Link>
-                </p>
-              )}
-            </>
-          )}
-        </Grid.Column>
-      </Grid>
-    </div>
+          </>
+        )}
+        {config.localUsers && (
+          <SignForm
+            submitText="Sign in"
+            handleSubmit={(e) => handleSubmit(e, formData, setFormData)}
+            formData={formData}
+            handleInputChange={handleInputChange}
+          />
+        )}
+      </Segment>
+      {config.hideSignup && (
+        <p>
+          If you do not have an account yet, please contact
+          <a href={`mailto:${config.adminEmail}`}> REANA administrators</a>
+        </p>
+      )}
+      {!config.hideSignup && config.localUsers && (
+        <p>
+          If you do not have an account yet, please
+          <Link to="/signup"> Sign up</Link> here
+        </p>
+      )}
+    </SignContainer>
   );
 }
-
-Signin.propTypes = {
-  /** Whether to show signup instead of signin form. */
-  signup: PropTypes.bool,
-};
-
-Signin.defatultProps = {
-  signup: false,
-};